refactor: migrate to gatsby v2; move some content around
