+++ content/speakers/eric-ponto/index.md
