+++ gatsby/create-pages.js
@@ -1,0 +1,32 @@
+const path = require('path');
+
+module.exports = ({graphql, actions}) => {
+  const {createPage} = actions;
+
+  return graphql(`
+      {
+        allMarkdownRemark {
+          edges {
+            node {
+              frontmatter {
+                date
+              }
+              fields {
+                slug
+              }
+            }
+          }
+        }
+      }
+    `).then(result => {
+    result.data.allMarkdownRemark.edges.forEach(({node}) => {
+      if (!node.fields || !node.fields.slug) return;
+
+      createPage({
+        path: node.fields.slug,
+        component: path.resolve('./src/templates/meeting.js'),
+        context: {slug: node.fields.slug}
+      });
+    });
+  });
+};
