+++ package.json
@@ -6,6 +6,9 @@
   "maintainers": [
     "Matt Travi <npm@travi.org> (https://matt.travi.org/)"
   ],
+  "contributors": [
+    "Dustin Schau <dustinschau@gmail.com> (https://dustinschau.com)"
+  ],
   "private": true,
   "repository": "dsmjs/site",
   "bugs": "https://github.com/dsmjs/site/issues",
@@ -16,6 +19,7 @@
     "test": "run-s lint:* build",
     "build": "gatsby build",
     "develop": "gatsby develop",
+    "serve": "gatsby serve",
     "start": "run-s develop",
     "commitmsg": "commitlint -e",
     "precommit": "npm test",
@@ -42,19 +46,20 @@
   },
   "dependencies": {
     "@dsmjs/components": "2.2.0",
-    "gatsby": "1.9.272",
-    "gatsby-link": "1.6.46",
-    "gatsby-plugin-glamor": "2.0.5",
+    "gatsby": "^2.0.0",
+    "gatsby-link": "^2.0.0",
+    "gatsby-plugin-glamor": "^2.0.0",
     "gatsby-plugin-google-fonts": "0.0.4",
-    "gatsby-plugin-react-helmet": "3.0.0",
-    "gatsby-remark-autolink-headers": "1.4.19",
-    "gatsby-remark-prismjs": "3.0.1",
-    "gatsby-source-filesystem": "1.5.39",
-    "gatsby-transformer-remark": "1.7.44",
-    "prismjs": "1.15.0",
-    "prop-types": "15.6.2",
-    "react": "16.5.2",
-    "react-dom": "16.5.2",
-    "react-helmet": "5.2.0"
+    "gatsby-plugin-react-helmet": "^3.0.0",
+    "gatsby-remark-autolink-headers": "^2.0.0",
+    "gatsby-remark-prismjs": "^3.0.1",
+    "gatsby-source-filesystem": "^2.0.0",
+    "gatsby-transformer-remark": "^2.0.0",
+    "glamor": "2.20.40",
+    "prismjs": "^1.15.0",
+    "prop-types": "^15.6.2",
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
+    "react-helmet": "^5.2.0"
   }
 }
