+++ src/.eslintrc.yml
@@ -2,3 +2,6 @@ extends: '@travi/travi/rules/react'
 
 globals:
   graphql: true
+
+rules:
+  react/jsx-no-bind: off # render props... ?
